pyproject.toml
[project]
name = "pico-agent"
version = "0.1.0"
description = "A pico-sized coding agent"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]
